Fix Space Objects Response
