+++ src/main/java/com/riwi/sitekeeper/services/SpaceService.java
@@ -7,12 +7,14 @@ import com.riwi.sitekeeper.dtos.requests.SpaceReq;
 import com.riwi.sitekeeper.dtos.responses.SpaceRes;
 import com.riwi.sitekeeper.entitites.SpaceEntity;
 import com.riwi.sitekeeper.repositories.SpaceRepository;
+import com.riwi.sitekeeper.utils.TransformUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class SpaceService {
@@ -23,18 +25,20 @@ public class SpaceService {
     @Autowired
     private NestServiceClient nestServiceClient;
 
+    private final TransformUtil transformUtil = new TransformUtil();
+
     public List<SpaceRes> getAllSpaces(String token) {
         List<SpaceEntity> spaces = spaceRepository.findAllByIsDeletedFalse();
         List<SpaceRes> spaceResList = new ArrayList<>();
         for (SpaceEntity space : spaces) {
-            spaceResList.add(convertToSpaceRes(space));
+            spaceResList.add(transformUtil.convertToSpaceRes(space));
         }
         return spaceResList;
     }
 
     public Optional<SpaceRes> getSpaceById(Long id, String token) {
         Optional<SpaceEntity> spaceOptional = spaceRepository.findById(id);
-        return spaceOptional.map(this::convertToSpaceRes);
+        return spaceOptional.map(transformUtil::convertToSpaceRes);
     }
 
     public Optional<SpaceEntity> getSpaceById(Long id) {
@@ -43,17 +47,17 @@ public class SpaceService {
 
     public Optional<SpaceRes> getSpaceByName(String name, String token) {
         Optional<SpaceEntity> spaceOptional = spaceRepository.findByName(name);
-        return spaceOptional.map(this::convertToSpaceRes);
+        return spaceOptional.map(transformUtil::convertToSpaceRes);
     }
 
     public SpaceRes createSpace(SpaceReq space, String token) {
         ValidationReq validationReq = new ValidationReq("spaces", "can_create");
         ValidationUserRes user = nestServiceClient.checkPermission(validationReq, token);
-        SpaceEntity newSpace = convertToSpaceEntity(space);
+        SpaceEntity newSpace = transformUtil.convertToSpaceEntity(space);
         newSpace.setCreatedBy(user.getId());
         newSpace.setUpdatedBy(user.getId());
         SpaceEntity savedSpace = spaceRepository.save(newSpace);
-        return convertToSpaceRes(savedSpace);
+        return transformUtil.convertToSpaceRes(savedSpace);
     }
 
     public SpaceRes updateSpace(Long id, SpaceReq updatedSpace, String token) {
@@ -61,9 +65,9 @@ public class SpaceService {
 
         if (existingSpaceOptional.isPresent()) {
             SpaceEntity existingSpace = existingSpaceOptional.get();
-            updateSpaceEntity(existingSpace, updatedSpace);
+            transformUtil.updateSpaceEntity(existingSpace, updatedSpace);
             SpaceEntity savedSpace = spaceRepository.save(existingSpace);
-            return convertToSpaceRes(savedSpace);
+            return transformUtil.convertToSpaceRes(savedSpace);
         } else {
             throw new RuntimeException("Space not found with id: " + id);
         }
@@ -71,32 +75,5 @@ public class SpaceService {
 
     public void deleteSpace(Long id, String token) {
         spaceRepository.softDeleteById(id);
-    }
-
-    public SpaceEntity convertToSpaceEntity(SpaceReq spaceReq) {
-        return SpaceEntity.builder()
-                .name(spaceReq.getName())
-                .location(spaceReq.getLocation())
-                .description(spaceReq.getDescription())
-                .image(spaceReq.getImage())
-                .build();
-    }
-
-    private SpaceRes convertToSpaceRes(SpaceEntity spaceEntity) {
-        return SpaceRes.builder()
-                .id(spaceEntity.getId())
-                .name(spaceEntity.getName())
-                .location(spaceEntity.getLocation())
-                .description(spaceEntity.getDescription())
-                .image(spaceEntity.getImage())
-                .objects(spaceEntity.getObjects())
-                .build();
-    }
-
-    private void updateSpaceEntity(SpaceEntity existingSpace, SpaceReq updatedSpace) {
-        existingSpace.setName(updatedSpace.getName());
-        existingSpace.setLocation(updatedSpace.getLocation());
-        existingSpace.setDescription(updatedSpace.getDescription());
-        existingSpace.setImage(updatedSpace.getImage());
     }
 }
